requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
hypothesis>=6.0.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.0

# Notebooks
jupyter>=1.0.0
ipython>=8.0.0

# Type stubs
pandas-stubs>=1.5.0
